Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/security/001-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/wrong_accept_key' failed: Error during WebSocket handshake: Sec-WebSocket-Accept mismatch

PASS WebSockets: wrong accept key 
